requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
openpyxl==3.1.5
pyyaml==6.0.2
matplotlib==3.9.2
joblib==1.4.2